health_modular_app/requirements.txt
Flask
Flask-MySQLdb
bcrypt
PyJWT
mysql-connector-python
